seo_agent/requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.1

# Google Ads
google-ads>=21.1.0

# Web Framework
fastapi>=0.95.0
uvicorn>=0.21.1

# Web and Data
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Async
httpx>=0.24.0
asyncio>=3.4.3

# Utils
tqdm>=4.65.0
python-slugify>=8.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
